requirements.txt
Django~=5.0.4
docxtpl~=0.16.8
openpyxl~=3.1.2
psycopg2-binary==2.9.3
python-dotenv~=1.0.1
gunicorn==20.1.0
djangorestframework==3.15.1
djoser==2.2.2